objs/AttackDisc: Implement a throwable attack
Implement a throwable attack, which is tightly couple to the entity that
threw it.

The entity that wants to use this must implement an interface so the
attack object may report events that happened while it was in effect.

The attack also linearly scales based on the attack power: the greater
the power, the bigger it is. Inversely, the bigger the attack, the shorter
its flight distance, and it's also slower.

Lastly, after an attack is thrown and it has fallen, it must be retrieved.
If it's not recovered within a time limit (inversely proportional to the
throw power), the attacking entity will be damaged.