objs/Player: Implement dash attack
Implement dash attack, a specific type of attack that heals the player
when successful, but damages it otherwise.

Since the attack object belongs to the Playstate, the state implements an
interface used by the player to request the attack to be spawned. This way,
if the attack object were moved elsewhere, the player wouldn't need to
change. However, this change would still be reflected in the spawner passed
to the player on its instantiation.